LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.baseline.default.worker-expected.txt


FAIL 2d assert_equals: ctx.textBaseline === 'alphabetic' (got [undefined], expected alphabetic[string]) expected (string) "alphabetic" but got (undefined) undefined
